Create chap4_exercise6.lean
It seems that we cannot rewrite a part of expression without tactics. I tried rewriting the proof without tactics, but there is an error messages that I have to provide a proof that the entire LHS equals RHS.

For example, it is not enough to provide `exp (log x) = x` to show `exp (log x) + y = x + y`, if we don't rely on tactics.